massive refactoring (see detailed description)
The factory methods in Domain (newBooleanValue, newFloatValue, newObject,...) are renamed and now have appropriate names starting with capital letters (BooleanValue, FloatValue, NewObject,...) to signify their purpose.

Moved the remaining concrete implementation of bytecode instructions (monitor enter and monitor exit) to a new trait.

Refined the return types (defined new type aliases) of the instructions to better communicate the expected result.

Added "override" to all methods and variable/values when possible. This facilitates comprehension and maintenance.

Added an extensive comment to AnIntegerValue to document when and how instances are/need to be created.

Removed some unused and also undesired factory methods.

Signed-off-by: Michael Eichberg <mail@michael-eichberg.de>