various refactoring to make the code base more modular, easier to handle and easier to comprehend
The trait "RecordReturnValues" was split up and the functionality is now represented by the traits "RecordLastReturnValues", "RecordAllThrownExceptions" and "RecordReturnInstructions" (The new implementation is now faster, more reusable and stable)

The type PCs is now defined in the package ai.

ClassFile now has a method "findMethod" to find some method with a specific name.

Signed-off-by: Michael Eichberg <mail@michael-eichberg.de>