Prevent Removal of Live Code in certain situations
The issue here is that the PCLabels are eagerly marked as live,
This then leads to a problem when a ControlTransfer Instruction is
marked as live after the corresponding Label has already been marked as
live. In that case the real instructions following that label are not
marked as live.
This commit explicitly searches forwards from PCLabels that are jump
targets to the first real instruction and marks it as live.